+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
+### WIP
 - Bootstrap Vue
 - header styles
 - Export to JSON and LEO format
@@ -14,6 +14,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - format @doc directive
 - fix to inline formatting (@language htmlsource, text)
 - fix to treeviewer scroll on windows
+
+## [1.12.6] - 2018-09-21
+### Changed
+- Fix to md IMG style
+
+## [1.12.5] - 2018-09-21
+### Added
+- Style improvements
+- lconfig options for node caret, leftPaneWidth, contentPaneWidth, highlightColor
+
+## [1.12.4] - 2018-09-11
+### Changed
+- Patch to fix style issues
+
+## [1.12.3] - 2018-09-11
+### Changed
+- Patch to fix directive highlighting
 
 ## [1.12.0] - 2018-09-08
 ### Added